refactored for reuse and decoupling
