+++ src/main/controllers/RespondToNotificationSubmitController.ts
@@ -2,7 +2,7 @@ import { Response } from 'express';
 
 import { AppRequest } from '../definitions/appRequest';
 import { ErrorPages, PageUrls, TseErrors } from '../definitions/constants';
-import { loadUserCaseFromApi } from '../helpers/CaseTransferHelper';
+import { LoadUserCaseResults, loadUserCaseFromApi } from '../helpers/LoadUserCaseHelper';
 import { getLanguageParam } from '../helpers/RouterHelpers';
 import { clearTempFields } from '../helpers/controller/RespondToNotificationSubmitHelper';
 import { getLogger } from '../logger';
@@ -31,11 +31,11 @@ export default class RespondToNotificationSubmitController {
         RespondentUtils.findSelectedRespondentByRequest(req)?.ccdId
       );
 
-      if (loadResult === 'transferred') {
+      if (loadResult === LoadUserCaseResults.TRANSFERRED) {
         return;
       }
 
-      if (loadResult === 'failed') {
+      if (loadResult === LoadUserCaseResults.FAILED) {
         return res.redirect(ErrorPages.NOT_FOUND + languageParam);
       }
